requirements.txt
ansi==0.1.3
beautifulsoup4==4.5.3
bottle==0.12.13
cffi==1.9.1
colorlog==2.10.0
cryptography==1.7.1
daemonize==2.4.7
dnspython==1.15.0
dnspython3==1.15.0
-e git+https://github.com/errbotio/errbot.git@5cdc0d3e883f55bdbcbc3408ac7d9b1bd352c094#egg=errbot
idna==2.2
inflect==0.2.5
-e git+https://github.com/jaraco/irc.git@8dbe54b87dc91dddaddd0c239041975fe189fccb#egg=irc
jaraco.classes==1.4
jaraco.collections==1.5
jaraco.functools==1.15.1
jaraco.itertools==2.0
jaraco.logging==1.5
jaraco.stream==1.1.1
jaraco.text==1.8
Jinja2==2.9.4
Markdown==2.6.7
MarkupSafe==0.23
more-itertools==2.5.0
pkg-resources==0.0.0
pyasn1==0.1.9
pycparser==2.17
Pygments==2.1.3
pygments-markdown-lexer==0.1.0.dev39
pyOpenSSL==16.2.0
pytz==2016.10
remote-pdb==1.2.0
requests==2.12.4
rocket-errbot==1.2.5
six==1.10.0
tempora==1.6.1
threadpool==1.3.2
waitress==1.0.1
WebOb==1.7.0
WebTest==2.0.24
Yapsy==1.11.223
